requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
Django==3.1.1
djangorestframework==3.11.1
idna==2.10
pytz==2020.4
requests==2.25.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
